% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedshear-package.R
\docType{package}
\name{bedshear-package}
\alias{bedshear}
\alias{bedshear-package}
\title{bedshear: Calculations for bed shear stress and wave orbital velocity}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
