% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shear_stress.R
\name{shear_stress}
\alias{shear_stress}
\title{Bed shear stress calculation}
\usage{
shear_stress(bathymetry = NULL, D50 = NULL, tidal_velocity = NULL,
  tidal_direction = NULL, wave_height = NULL, wave_period = NULL,
  wave_direction = NULL, switch = 0)
}
\arguments{
\item{bathymetry}{Bathymetry of location}

\item{D50}{Median grain size (mm).}

\item{tidal_velocity}{A numeric vector of tidal velocity (m/s).}

\item{tidal_direction}{A numeric vector of tidal directions (degrees).}

\item{wave_height}{Significant wave height (m).}

\item{wave_period}{Wave period.}

\item{wave_direction}{Wave direction.}

\item{switch}{Switching term. Defaults to 0.}
}
\value{
data frame with bed shear stress etc.
}
\description{
Calculates the combined bed shear stress from waves and tides
}
\examples{
# Calculating a time series of bed shear stress at Stonehaven
library(bedshear)
library(dplyr)
library(ggplot2)

stonehaven_stress <- shear_stress(bathymetry = 40,0.0002, stonehaven_ts$tidal_velocity, stonehaven_ts$tidal_direction,stonehaven_ts$wave_height,
																		stonehaven_ts$wave_period , switch = 0, stonehaven_ts$wave_direction )

stonehaven_ts \%>\%
	mutate(Stress = stonehaven_stress$shearmax) \%>\%
	ggplot(aes(Date, Stress))+
geom_line()
}
