% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shear_stress.R
\name{shields_bedshear}
\alias{shields_bedshear}
\title{Bed shear stress calculation}
\usage{
shields_bedshear(bathymetry = NULL, D50 = NULL,
  tidal_velocity = NULL, tidal_direction = NULL, wave_height = NULL,
  wave_period = NULL, wave_direction = NULL, switch = 0)
}
\arguments{
\item{bathymetry}{Bathymetry of location}

\item{D50}{Median grain size (mm).}

\item{tidal_velocity}{A numeric vector of tidal velocity (m/s).}

\item{tidal_direction}{A numeric vector of tidal directions (degrees).}

\item{wave_height}{Significant wave height (m).}

\item{wave_period}{Wave period.}

\item{wave_direction}{Wave direction.}

\item{switch}{Switching term. Defaults to 0.}
}
\value{
data frame with bed shear stress etc.
}
\description{
Calculates the combined bed shear stress from waves and tides
}
\examples{
# Calculating cell areas for NOAA world ocean atlas data sample file
ff <- system.file("extdata", "woa18_decav_t01_01.nc", package = "rcdo")
nc_cellareas(ff)

# Calculating a time series of bed shear stress at Stonehaven
library(bedshear)
library(dplyr)
library(ggplot2)
stonehaven_stress <- shields_bedshear(Depth,0.0002, sample_data$tidal_velocity, sample_data$tidal_direction,sample_data$wave_height,
																		sample_data$wave_period , switch = 0, sample_data$wave_direction )
stonehaven_ts \%>\%
	mutate(Stress = new_result$shearmax) \%>\%
	ggplot(aes(Date, Stress))+
	geom_line()
}
