% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stonehaven_ts.R
\docType{data}
\name{stonehaven_ts}
\alias{stonehaven_ts}
\title{Wave and tidal data for a site near Stonehaven, Scotland}
\format{A data frame with 35040 rows and 8 variables
\describe{
  \item{tidal_velocity}{depth averaged tidal velocity, in m/s}
  \item{tidal_direction}{tidal direction}
  \item{wave_height}{significant wave height, in m}
  \item{wave_period}{wave period}
  \item{wave_direction}{wave direction}
}}
\source{
\url{https://www.sciencedirect.com/science/article/pii/S0964569116302587}
}
\usage{
stonehaven_ts
}
\description{
A time series of wave and tidal data for a station near Stonehaven, Scotland.
Depth of this site is 38 metres and the D50 is 0.2 mm
Reference: https://www.sciencedirect.com/science/article/pii/S0964569116302587
}
\keyword{datasets}
