% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shear_stress.R
\name{shear_stress}
\alias{shear_stress}
\title{Bed shear stress calculation}
\usage{
shear_stress(bathymetry = NULL, D50 = NULL, tidal_velocity = NULL,
  tidal_direction = NULL, wave_height = NULL, wave_period = NULL,
  wave_direction = NULL, switch = 0)
}
\arguments{
\item{bathymetry}{Bathymetry of location}

\item{D50}{Median grain size (mm).}

\item{tidal_velocity}{A numeric vector of tidal velocity (m/s). If not given, this is set to zero.}

\item{tidal_direction}{A numeric vector of tidal directions (degrees). If not given, this is set to zero.}

\item{wave_height}{Significant wave height (m). If not given, this is set to zero.}

\item{wave_period}{Wave period. If not given, this is set to zero.}

\item{wave_direction}{Wave direction. If not given, this is set to zero.}

\item{switch}{Switching term. Defaults to 0.}
}
\value{
Data frame with mean shear stress, maximum shear stress, root mean square shear stress, tidal current only shear stress (if tides supplied),
wave-only shear stress (if waves supplied), Shields number, Critical Shields,  Flow type.
}
\description{
Calculates the combined bed shear stress from waves and tides
}
\examples{

# Calculating a time series of bed shear stress at Stonehaven
library(bedshear)
library(dplyr)
library(ggplot2)

stonehaven_stress <- shear_stress(
  bathymetry = 40, D50 = 0.0002, tidal_velocity = stonehaven_ts$tidal_velocity, tidal_direction = stonehaven_ts$tidal_direction, wave_height = stonehaven_ts$wave_height,
  wave_period = stonehaven_ts$wave_period, wave_direction = stonehaven_ts$wave_direction, switch = 0
)

stonehaven_ts \%>\%
  mutate(Stress = stonehaven_stress$shear_max) \%>\%
  ggplot(aes(Date, Stress)) +
  geom_line()
}
